Create add new template modal.
