+++ src/pages/Templates/_components/index.ts
@@ -1,0 +1,1 @@
+export * from "./AddTemplate"
